refactoring activities cache for kadena

+++ packages/extension/src/libs/activity-state/wrap-activity-handler.ts
@@ -2,6 +2,7 @@ import { ActivityStatus } from "@/types/activity";
 import ActivityState from ".";
 import { ActivityHandlerType } from "./types";
 const CACHE_TTL = 1000 * 60 * 5; // 5 mins
+import { ProviderName } from "@/types/provider";
 export default (activityHandler: ActivityHandlerType): ActivityHandlerType => {
   const returnFunction: ActivityHandlerType = async (network, address) => {
     const activityState = new ActivityState();
@@ -13,35 +14,36 @@ export default (activityHandler: ActivityHandlerType): ActivityHandlerType => {
       activityState.getAllActivities(options),
       activityState.getCacheTime(options),
     ]);
-    console.log("activities from cache", activities);
-    if (cacheTime + CACHE_TTL < new Date().getTime()) {
-      console.log("OUTSIDE cache time");
-      const liveActivities = await activityHandler(network, address);
-      const activitiesToAdd = activities.filter(
-        (a) =>
-          a.status !== ActivityStatus.success &&
-          a.status !== ActivityStatus.failed
-      );
-      await activityState.deleteAllActivities(options);
-      await activityState.addActivities(
-        liveActivities.concat(activitiesToAdd),
-        options
-      );
-      await activityState.setCacheTime(options);
-      return activityState.getAllActivities(options);
-
-      // if (!activities.length) {
-      //   await activityState.addActivities(liveActivities, options);
-      //   await activityState.setCacheTime(options);
-      //   return liveActivities;
-      // } else {
-      //   await activityState.addActivities(liveActivities, options);
-      //   await activityState.setCacheTime(options);
-      //   return activityState.getAllActivities(options);
-      // }
+    const liveActivities = await activityHandler(network, address);
+    if (network.provider === ProviderName.kadena) {
+      if (cacheTime + CACHE_TTL < new Date().getTime()) {
+        const inProcessActivities = activities.filter(
+          (a) =>
+            a.status !== ActivityStatus.success &&
+            a.status !== ActivityStatus.failed
+        );
+        await activityState.deleteAllActivities(options);
+        await activityState.addActivities(
+          liveActivities.concat(inProcessActivities),
+          options
+        );
+        await activityState.setCacheTime(options);
+        return activityState.getAllActivities(options);
+      } else return activities;
     } else {
-      console.log("INSIDE cache time");
-      return activities;
+      if (cacheTime + CACHE_TTL < new Date().getTime()) {
+        if (!activities.length) {
+          await activityState.addActivities(liveActivities, options);
+          await activityState.setCacheTime(options);
+          return liveActivities;
+        } else {
+          await activityState.addActivities(liveActivities, options);
+          await activityState.setCacheTime(options);
+          return activityState.getAllActivities(options);
+        }
+      } else {
+        return activities;
+      }
     }
   };
   return returnFunction;
